Move upgrade-ttl as a sub command of ledis so that it's easier to repair the problems
